Limpeza no código
Realizado limpeza dos dados e comentários inúteis/desnecessários do código.

+++ src/main/java/Product.java
@@ -1,10 +1,10 @@
 public class Product {
 
-    private final Long id;
+    private final Double id;
     private final Double price;
 
     public Product(String[] lineData) {
-        this.id = Long.valueOf(lineData[0]);
+        this.id = Double.valueOf(lineData[0]);
         this.price = Double.valueOf((lineData[2]));
     }
